Login and stuff
I'm sorry for the very long descriptions.
As I said in the previous update, I needed to test whether or not users could login.
Tested it, and they couldn't. Fixed it.
There is no password encryption yet, but that can be done in like... five minutes, so that's *probably* the next thing I'm going to do. That and a bit of... beautification for the login page.
Also worked a tiny bit on the forum structure, I guess that's what I should call it.
And before I forget to mention everything is inaccessible if the user doesn't log in, this is by design not a bug, I just need to make it a bit more visually apparent.
But for now (in other words "today") I'm done, did not get a lot of sleep.