Registration and a bit of everthing
Registration works now. Added a new table "category" for the forum page... currently empty. Worked a tiny bit on the layout of the pages. Users now cannot access any page outside of the portal, login and registration page unless they're logged in. There is now the capability in the code to switch between the two registration modes (student and teacher). Teacher registration is currently a blank page. Password encryption is done.

Issue with the registration:
If the user refreshes the page after a successful registartion it registers another account with the same account details(email username password,etc.)

To Do:
- Fix afformentioned issue with the registration page.
- Create layout for blank pages(such as the profile page).