Registration complete
Now users can register and the system checks the email address, username and password.
Email address:
Checks if it's already in use.
Checks if it's an actual address (not the best but it works).

Username:
For now it only checks if it's already in use. I'll probably set up a blacklist of words, such as profanity.

Password:
Checks if its at least 8 characters long.
Checks if it contains at least one letters.
Checks if it contains at least one uppercase letter.
Checks if it contains at least one number.

For now I think the registration and login part of the site is done from a coding point of view. Visual design is another question. but that can wait.
It's easier to beautify a complete codes visuals then the other way around.

Next step:
Profile page (code and layout).