Profile Picture upload
Users can now upload a profile picture in the Profile page. Said profile picture is then stored locally.
I put the upload into a switch because there will be a change option as well.
It doesn't display the profile pic yet.
but it does upload it without any issues, so I'm satisfied for now.