Registration once again
YEEHAW FELLAS!!!!

IT WORKS!!!

IT'S ALIVE!!!

IT *BREATHES*

*clears throat*

So yes, registration is once again completely functional.
Also up until now I had the wrong sql file here on github, it's the correct one now and I moved it out of the assets folder for... reasons.
And I also got rid of (I think) every bit of unused code and simplified stuff where I could as it was overcomplicated for no fucking reason, so it may a bit more comprehensive now.
Right and the login wasn't checking for the username correctly.
Doesn't work on phone currently... and it'll stay like that for a while too as it's not my priority. I had enough for today as it was incredibly frustrating.

But right now I'm in bliss.
And at last there is Holy Silence.

...

It's a good song, go listen to it.