Worked on the chat page
I worked on the css a little bit.
also made a new model file for the chat
reworked the database to be compatible with a group chat.
For some reason some old files were still in this git, so I deleted them.
Also there is now a notifications button... which currently does nothing.