Fffucking front-end
I hate it.
I despise it.
And I cannot run from it.
So the best I can do is contain my anger and try.

Anyway.
The login and register pages are now responsive, so the site is now usable on phone.
Also the login and register pages now look... "decent".

The chat page got fucked.
I honestly don't know why.
Like I have no clue and I've spent the last hour trying to figure out to no avail. So that's it for today.

Also from this point on I will recommend a song at the end of every commit I push just for the fuck of it, in case anyone is reading these. So with that said...

Listen to Mon Amour, Mon Ami by Therion. Hope you like the french.
*I* sure as hell don't.