Front-end once again
AAAAAAAAAAAAAAAAAAAH

Anyway, made a lot of progress on the look of the chat page, dare I say it's even starting to look good.
Also made some adjustments on the navbar, not much worth mentoining any of it.

Song:
Romanticide
Band:
Nightwish

It's goth, its gooooooooooooood.

That's all for today cause I have a head ache.