IT LIVES
The chat works!

It's ALIVE, *BREATHING*.

There is a pulse.

YEEEEEEEEEEEEEES!

All it took was 60% of my remaining sanity (wasn't much to begin with), a pact with the Samigina of the Goetia and a *lot* of screaming.

Anyway, so the chat is almost completely functional now (can't create chatrooms yet), it just isnt live yet. I'll leave that for later, next step is the *purpose* of the entire page, the forum.

Song: Ready To Die
Artist: Andrew W. K.