Chat fixes
There were 2 issues witht the chat.
First the method I used to determine the  id of messages was subpar and cause an sql warning for every single message in the database one by one. This has now been fixed.
Second when a user would refresh the page after sending a message the site would resend the message upon the refresh. This has been... semi-fixed. Or rather it's a temporary fix using javascript, this issue will be completely gotten rid of once I implement the ajax part of the site.

Song: Acheron
Artist: Two Feather and Randy Blythe of Lamb of God