A bit of everything
When registering a new user it encrypts the user's email address.
There is now a new button in the chat page at the end of the room:
Users will be able to create their own *private* rooms where they have to invite other users to talk to manually.
All public rooms are owned by the site owner.
The button opens up a modal where the room will be created.
The function is done for it. The front-end is mostly done for it. I just need to make it do the deed when the button is pressed.
But my internet is fading in and out of existance today so I have to stop here as I'm literally unable to work.
Also just worked on the chat page's front-end again. No aneurysm this time though.

Song: Drown
Artist: Kim Dracula
Full of Hatred and Wrath. Is good.