Private Rooms Work
Yeehaw Chucklefucks,
Users can now create private rooms, there is a limit of 1 private room per user. The owner of the room can invite other users into the private room.
For the private room to work I had to modify the database, but this was very minor, I just added a table that is used for determining who owns what room who is just a member of said room.
Smaller front-end changes on the chat page.

The login pages now redirects the user to the main page upon a successful login. The regsitration page does the same and also logs the user in.
This was done becasue *SOMEONE* wouldn't stop bothering me about it, as if that was the biggest priority, so here you go Tokris, you fucking bastard.
Another change is that the Forum page is now the site's main page, and users now have to login the moment they open the site.
Also as a last note a larger amount of the already existing functions in the chat model have been modified to a good degree. Only mentioning it cause a lot of them were changed, so I think it's worth mentioning.

Song: Enemy Within
Artist: Arch Enemy