THE CHAT IS LIVE
IT WAS 3 LINES OF CODE.
So the chat is now live, it needs some refinement, such as accepting special characters, but FUCK YEAH, it works.
Furthermore the forum page is close to completion in terms of displaying things.
I decided to make my own modals instead of bootstrap ones. I'm halfway through with making those, one works completely, the second one is smei working, as in the form inside it does nothing yet, and the third one still uses the bootstrap one.
The site now uses session_destroy instead of unset at Kori's suggestion.
Also umm a mistake was made a while ago I maaaaaaaay have forgotten to *PROPERLY* set up the connections between the tables, they were all in restrict, now in cascade. I'll fix most issue with the chat tomorrow and I think I'll leave the rest of the forum page for the rest of the week, mostly the weekend, and I'll do the profile page next week. After that... I think I'm done after that, just beautifying and thats it, also the documentation but I don't have to commit that.

Song: Emerald Sword
Artist: Rhapsody of Fire