Forum and chat page
Chat page had small adjustment on its front-end and back-end equally.
Users can now upload a post to the database, they can't upload a picture with them yet.
The page doesn't output the post correctly for now.
The goal for today is to fix that and the picture upload.

Also thanks horvath we got rid of an issue at has arisen from the fact that we use a shared server, that being that the session would mix up when someone would visit someone else's site.
Also also the profile picture upload doesn't work atm as I'm going to remake it's front-end so that it's one button.

Song: New Horizon (I Found You)
Artist: Outrage