Front-end Again
Chat page is almost done honestly with in every term, the back-end is completely implamanted, if i do anything it'll be minor adjustments.
The front-end for it is almost done.
Forum page is going well too, the back-end has no issues as of yet, I need to make it presentable, and I need to implement QoL features for it along side the comment section, but honestly that's going to be pretty easy.
Profile page is empty and has no fucntion atm, Admin page doesn't exit yet. Will do soon.

Song: Crawl
Artists: Daisuke Ishiwatari and Naoki Hashimoto