Front-end 9: The Reckoning
It looks alright on phone now, the front-end for the forum page is almost done, at least for the part that actually works.

Also in the previous commit I forgot to mention that I had to modify the database. I modified the reaction table so that it only handles the reactions for posts and I added two more tables that do the same for the comments.
It should work once I do it.

Song: I'll Make Them Whole
Artist Will Ryan - DAGames