3 down 4 more to go
(I'm only talking about the back-end)
3 Pages are done and completely working (Chat, login, register).
I'm half-way through with the forum page, and the profile is easy, the admin page won't be too hard either, and the last one is the page for the individual posts, that will be a bit more difficult but nothing I can't do. After all of that, all that will remain is making it look normal on mobile (in other words the bad part) and the documentation the worse part.

Song: Hate/Love
Artist: Electric Callboy